API Documentation

OptimalTransportNetworks.add_node

OptimalTransportNetworks.annealing

OptimalTransportNetworks.apply_geography

OptimalTransportNetworks.create_graph

OptimalTransportNetworks.find_node

OptimalTransportNetworks.init_parameters

OptimalTransportNetworks.optimal_network

OptimalTransportNetworks.plot_graph

OptimalTransportNetworks.remove_node

OptimalTransportNetworks.add_node — Method

add_node(param, graph, x, y, neighbors) -> Dict, NamedTuple

Add a node in position (x,y) and list of neighbors. The new node is given an index J+1.

Arguments

param::Dict: Dict that contains the model's parameters, or nothing to only update graph

graph::NamedTuple: Named tuple that contains the underlying graph (created by create_graph())

x::Float64: x coordinate of the new node (any real number)

y::Float64: y coordinate of the new node (any real number)

neighbors::Vector{Int64}: Vector of nodes to which it is connected (1 x n list of node indices between 1 and J, where n is an arbitrary # of neighbors)

The cost matrices delta_tau and delta_i are parametrized as a function of Euclidean distance between nodes.

Returns the updated graph and param objects (param is affected too because the variable Zjn, Lj, Hj and others are reset to a uniform dist.)

OptimalTransportNetworks.annealing — Method

annealing(param, graph, I0; kwargs...)

Runs the simulated annealing method starting from network I0. Only sensible if param[:gamma] > param[:beta].

Arguments

param: Dict that contains the model's parameters

graph: Named tuple that contains the underlying graph

I0: (optional) provides the initial guess for the iterations

Keyword Arguments

perturbation_method::String="random rebranching": Method to be used to perturbate the network (random is purely random, works horribly; shake applies a gaussian blur along a random direction, works alright; rebranching (default) is the algorithm described in Appendix A.4 in the paper, works nicely )

preserve_central_symmetry::Bool=false: Only applies to shake method

preserve_vertical_symmetry::Bool=false: Only applies to shake method

preserve_horizontal_symmetry::Bool=false: Only applies to shake method

smoothing_radius::Float64=0.25: Parameters of the Gaussian blur

mu_perturbation::Float64=log(0.3): Parameters of the Gaussian blur

sigma_perturbation::Float64=0.05: Parameters of the Gaussian blur

display::Bool: Display the graph in each iteration as we go

t_start::Float64=100: Initial temperature

t_end::Float64=1: Final temperature

t_step::Float64=0.9: Speed of cooling

num_deepening::Int64=4: Number of FOC iterations between candidate draws

num_random_perturbations::Int64=1: Number of links to be randomly affected ('random' and 'random rebranching' only)

Iu::Matrix{Float64}=Inf * ones(J, J): J x J matrix of upper bounds on network infrastructure Ijk

Il::Matrix{Float64}=zeros(J, J): J x J matrix of lower bounds on network infrastructure Ijk

model::Function: For custom models => a function that taks an optimizer and an 'auxdata' structure as created by create_auxdata() as input and returns a fully parameterized JuMP model

final_model::JuMPModel: Alternatively: a readily parameterized JuMP model to be used (from optimal_network())

recover_allocation::Function: The recover_allocation() function corresponding to either model or final_model

Examples

# Nonconvex case, disabling automatic annealing
param = init_parameters(annealing = false, gamma = 2)
param, graph = create_graph(param)
param[:Zjn][61] = 10.0
result = optimal_network(param, graph)

# Run annealing
results_annealing = annealing(param, graph, result[:Ijk])

# Comparison
plot_graph(graph, result[:Ijk])
plot_graph(graph, result_annealing[:Ijk])

OptimalTransportNetworks.apply_geography — Method

apply_geography(graph, geography; kwargs...) -> updated_graph

Update the network building costs of a graph based on geographical features and remove edges impeded by geographical barriers. Aversion to altitude changes rescales building infrastructure costs delta_i by (see also user manual to MATLAB toolbox):

1 + alpha_up * max(0, Z2-Z1)^beta_up + alpha_down * max(0, Z1-Z2)^beta_down

and similarly for graph traversal costs delta_tau.

Arguments

graph::NamedTuple: The network graph to which the geographical features will be applied.

geography::NamedTuple: A named tuple representing the geographical features, with the following fields:

z::Vector{Float64}: A J x 1 vector containing the z-coordinate (elevation) for each node, or nothing if no elevation data.

z_is_friction::Bool: (Optional) logical value indicate that z represents friction rather than elevation. In that case, the measure of building cost is the average friction of the two nodes mean(Z1,Z2) rather than the difference Z2-Z1.

obstacles::Matrix{Int64}: An Nobs x 2 matrix specifying (i, j) pairs of nodes that are connected by obstacles, where Nobs is the number of obstacles, or nothing if no obstacles.

Keyword Arguments

across_obstacle_delta_i::Float64=Inf: Rescaling parameter for building cost that crosses an obstacle.

along_obstacle_delta_i::Float64=Inf: Rescaling parameter for building cost that goes along an obstacle.

across_obstacle_delta_tau::Float64=Inf: Rescaling parameter for transport cost that crosses an obstacle.

along_obstacle_delta_tau::Float64=Inf: Rescaling parameter for transport cost that goes along an obstacle.

alpha_up_i::Float64=0: Building cost scale parameter for roads that go up in elevation.

beta_up_i::Float64=1: Building cost elasticity parameter for roads that go up in elevation.

alpha_up_tau::Float64=0: Transport cost scale parameter for roads that go up in elevation.

beta_up_tau::Float64=1: Transport cost elasticity parameter for roads that go up in elevation.

alpha_down_i::Float64=0: Building cost scale parameter for roads that go down in elevation.

beta_down_i::Float64=1: Building cost elasticity parameter for roads that go down in elevation.

alpha_down_tau::Float64=0: Transport cost scale parameter for roads that go down in elevation.

beta_down_tau::Float64=1: Transport cost elasticity parameter for roads that go down in elevation.

Examples

param, graph = create_graph(init_parameters())
geography = (z = rand(graph.J), obstacles = [1 15; 70 72])
updated_graph = apply_geography(graph, geography)

OptimalTransportNetworks.create_graph — Function

create_graph(param, w = 11, h = 11; type = "map", kwargs...) -> Dict, NamedTuple

Initialize the underlying graph, population and productivity parameters.

Arguments

param::Dict: Structure that contains the model parameters

w::Int64=11: Number of nodes along the width of the underlying graph (integer)

h::Int64=11: Number of nodes along the height of the underlying graph (integer, odd if triangle)

Keyword Arguments

type::String="map": Either "map", "square", "triangle", or "custom"

omega::Vector{Float64}: Vector of Pareto weights for each node or region in partial mobility case (default ones(J or nregions))

Zjn::Matrix{Float64}: J x N matrix of producties per node (j = 1:J) and good (n = 1:N) (default ones(J, N))

adjacency::BitMatrix: J x J Adjacency matrix (only used for custom network)

x::Vector{Float64}: x coordinate (longitude) of each node (only used for custom network)

y::Vector{Float64}: y coordinate (latitude) of each node (only used for custom network)

nregions::Int64: Number of regions (only for partial mobility)

region::Vector{Int64}: Vector indicating region of each location (only for partial mobility)

Lj::Vector{Float64}: Vector of popultions in each node (j = 1:J) (only for no mobility)

Lr::Vector{Float64}: Vector of popultions in each region (r = 1:nregions) (only for partial mobility)

Hj::Vector{Float64}: Vector of immobile good in each node (j = 1:J) (e.g. housing, default ones(J))

Examples

param, graph = create_graph(init_parameters())

OptimalTransportNetworks.find_node — Method

find_node(graph, x, y) -> Int64

Returns the index of the node closest to the coordinates (x,y) on the graph.

Arguments

graph::NamedTuple: structure that contains the underlying graph

x::Float64: x coordinate on the graph (between 1 and w)

y::Float64: y coordinate on the graph (between 1 and h)

OptimalTransportNetworks.init_parameters — Method

init_parameters(; kwargs...) -> Dict

Returns a param dict with the model parameters.

Keyword Arguments

alpha::Float64=0.5: Cobb-Douglas coefficient on final good c^alpha * h^(1-alpha)

beta::Float64=1: Parameter governing congestion in transport cost

gamma::Float64=1: Elasticity of transport cost relative to infrastructure

K::Float64=1: Amount of concrete/asphalt

sigma::Float64=5: Elasticity of substitution across goods (CES)

rho::Float64=2: Curvature in utility (c^alpha * h^(1-alpha))^(1-rho)/(1-rho)

a::Float64=0.8: Curvature of the production function L^alpha

m::Vector{Float64}=ones(N,1): Vector of weights Nx1 in the cross congestion cost function

N::Int64=1: Number of goods

nu::Float64=1: Elasticity of substitution b/w goods in transport costs if cross-good congestion

labor_mobility::Any=false: Switch for labor mobility (true/false or 'partial')

cross_good_congestion::Bool=false: Switch for cross-good congestion

annealing::Bool=true: Switch for the use of annealing at the end of iterations (only if gamma > beta)

verbose::Bool=true: Switch to turn on/off text output (from Ipopt or other optimizers)

duality::Bool=true: Switch to turn on/off duality whenever available

warm_start::Bool=false: Use the previous solution as a warm start for the next iteration

kappa_tol::Float64=1e-7: Tolerance for convergence of road capacities κ

kappa_min::Float64=1e-5: Minimum value for road capacities κ

kappa_min_iter::Int64=20: Minimum number of iterations

kappa_max_iter::Int64=200: Maximum number of iterations

optimizer_attr::Dict: Dict of attributes passed to the optimizer (e.g. Dict(:tol => 1e-5))

model_attr::Dict: Dict of tuples (length 2) passed to the model (e.g. Dict(:backend => (MOI.AutomaticDifferentiationBackend(), MathOptSymbolicAD.DefaultBackend())) to use Symbolic AD)

model::Function: For custom models => a function that taks an optimizer and an 'auxdata' structure as created by create_auxdata() as input and returns a fully parameterized JuMP model

recover_allocation::Function: For custom models => a function that takes a solution and 'auxdata' structure as input and returns the allocation variables. In particular, it should return a dict with symbol keys returning at least objects :welfare => scalar welfare measure, :Pjn => prices, :PCj => aggregate condumption, and :Qjkn => flows.

Examples

param = init_parameters(labor_mobility = true, K = 10)

OptimalTransportNetworks.optimal_network — Method

optimal_network(param, graph; I0=nothing, Il=nothing, Iu=nothing, 
                verbose=false, return_model=0) -> Dict

Solve for the optimal network by solving the inner problem and the outer problem by iterating over the FOCs.

Arguments

param::Dict: Dict that contains the model's parameters

graph::NamedTuple: Named tuple that contains the underlying graph (created by create_graph() function)

I0::Matrix{Float64}: (Optional) J x J matrix providing the initial guess for the iterations

Il::Matrix{Float64}: (Optional) J x J matrix providing exogenous lower bound on infrastructure levels

Iu::Matrix{Float64}: (Optional) J x J matrix providing exogenous upper bound on infrastructure levels

verbose::Bool: (Optional) tell IPOPT to display results

return_model::Int=0: (Optional) return the JuMP model and corresponding recover_allocation() function: 1 just returns these before solving the model, while 2 solves the model + optimal network and returns the two alongside the results.

Examples

param = init_parameters()
param, graph = create_graph(param)
param[:Zjn][61] = 10.0
result = optimal_network(param, graph)
plot_graph(graph, result[:Ijk])

OptimalTransportNetworks.plot_graph — Function

plot_graph(graph, edges = nothing; kwargs...) -> Plots.Plot

Plot a graph visualization with various styling options.

Arguments

graph::NamedTuple: The network graph (created with create_graph())

edges::Matrix{Float64}=nothing: Matrix of edge weights (J x J)

Keyword Arguments

grid::Bool=false: Show gridlines

axis::Tuple=([], false): Axis ticks and labels (see Plots.jl docs, default disable axis)

margin::Real=-30mm: Margin around plot

map::Vector=nothing: Values mapped to graph for background heatmap

map_color::Symbol=:YlOrBr_4: Colorscale for background heatmap

mesh::Bool=false: Show mesh lines between nodes

mesh_color::Symbol=:grey90: Color for mesh lines

mesh_style::Symbol=:dash: Linestyle for mesh lines

mesh_transparency::Real=1: Opacity for mesh lines

edges::Bool=true: Show edges between nodes

edge_color::Symbol=:blue: Edge color or color gradient

edge_scaling::Bool=false: Size edges based on raw values

edge_transparency::Union{Bool,Real}=true: Transparency for edges

edge_min::Real: Minimum edge value for scaling

edge_max::Real: Maximum edge value for scaling

edge_min_thickness::Real=0.1: Minimum thickness for edges

edge_max_thickness::Real=2: Maximum thickness for edges

arrows::Bool=false: Show arrowheads on edges

arrow_scale::Real=1: Scaling factor for arrowheads

arrow_style::String="long": Style of arrowheads ("long" or "thin")

nodes::Bool=true: Show nodes

node_sizes::Vector=ones(J): Sizes for nodes

node_sizes_scale::Real=75: Overall scaling for node sizes

node_shades::Vector=nothing: Shades mapped to nodes

node_color::Symbol=:purple: Node color or color gradient

node_stroke_width::Real=0: Stroke width for node outlines

node_stroke_color::Symbol=nothing: Stroke color for node outlines

geography::NamedTuple=nothing: Named tuple with geography data, see also apply_geography()

obstacles::Bool=false: Show obstacles from geography

obstacle_color::Symbol=:black: Color for obstacles

obstacle_thickness::Symbol=3: Thickness for obstacles

Examples

param = init_parameters()
param, graph = create_graph(param)
param[:Zjn][51] = 10.0
result = optimal_network(param, graph)
plot_graph(graph, result[:Ijk])

OptimalTransportNetworks.remove_node — Method

remove_node(param, graph, i) -> updated_param, updated_graph

Removes node i from the graph.

Arguments

param::Dict: structure that contains the model's parameters, or nothing to only update graph

graph::NamedTuple: structure that contains the underlying graph (created by create_graph())

i::Int64: index of the mode to be removed (integer between 1 and graph.J)

Returns the updated graph and param objects (param is affected too because the variable Zjn, Lj, Hj and others are changed).